backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-nmap==0.7.1
pymongo==4.6.0
reportlab==4.0.7
